// src/typechain/Oracle.ts
export declare namespace IOracle {
  export type RoundDataStruct = {
    round: BigNumberish;
    price: BigNumberish;
    timestamp: BigNumberish;
    success: boolean;
  };

  export type RoundDataStructOutput = [
    round: bigint,
    price: bigint,
    timestamp: bigint,
    success: boolean
  ] & { round: bigint; price: bigint; timestamp: bigint; success: boolean };
}